# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List tracked appointments
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Track a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/services:
    post:
      operationId: adminCreateService
      tags: [admin]
      summary: Create a service (staff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /admin/services/{id}:
    patch:
      operationId: adminUpdateService
      tags: [admin]
      summary: Update a service (staff)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: adminDeleteService
      tags: [admin]
      summary: Delete a service (staff)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /admin/reward-items:
    get:
      operationId: adminListRewardItems
      tags: [admin]
      summary: List all reward items including inactive (staff)
      responses:
        "200":
          description: Reward items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RewardItem"
    post:
      operationId: adminCreateRewardItem
      tags: [admin]
      summary: Create a reward item (staff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardItemInput"
      responses:
        "201":
          description: Created reward item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardItem"
  /admin/reward-items/{id}:
    patch:
      operationId: adminUpdateRewardItem
      tags: [admin]
      summary: Update a reward item (staff)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardItemUpdate"
      responses:
        "200":
          description: Updated reward item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: adminDeleteRewardItem
      tags: [admin]
      summary: Deactivate a reward item (staff)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deactivated
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, name, category, description, bookingUrl]
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        description: { type: string }
        durationMinutes: { type: ["integer", "null"] }
        priceText: { type: ["string", "null"] }
        bookingUrl: { type: string }
...
    Appointment:
      type: object
      required: [id, serviceName, date, status]
      properties:
        id: { type: integer }
        serviceName: { type: string }
        providerName: { type: ["string", "null"] }
        date: { type: string }
        time: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        status: { type: string }
    AppointmentInput:
      type: object
      required: [serviceName, date]
      properties:
        serviceName: { type: string, minLength: 1 }
        providerName: { type: string }
        date: { type: string }
        time: { type: string }
        notes: { type: string }
        status: { type: string }
    AppointmentUpdate:
      type: object
      properties:
        serviceName: { type: string, minLength: 1 }
        providerName: { type: string }
        date: { type: string }
        time: { type: string }
        notes: { type: string }
        status: { type: string }
...
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ServiceInput:
      type: object
      required: [name, category, description]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        durationMinutes: { type: ["integer", "null"] }
        priceText: { type: ["string", "null"] }
        bookingUrl: { type: string }
    ServiceUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        durationMinutes: { type: ["integer", "null"] }
        priceText: { type: ["string", "null"] }
        bookingUrl: { type: string }
    RewardItem:
      type: object
      required: [id, title, description, points, active, sortOrder]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        points: { type: integer }
        active: { type: boolean }
        sortOrder: { type: integer }
    RewardItemInput:
      type: object
      required: [title, description, points]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        points: { type: integer, minimum: 1 }
        active: { type: boolean }
        sortOrder: { type: integer }
    RewardItemUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        points: { type: integer, minimum: 1 }
        active: { type: boolean }
        sortOrder: { type: integer }
...