# lib/api-zod/src/generated/types/rewardItemInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RewardItemInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  /** @minimum 1 */
  points: number;
  active?: boolean;
  sortOrder?: number;
}