# lib/api-zod/src/generated/types/serviceUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  /** @minLength 1 */
  description?: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  priceText?: string | null;
  bookingUrl?: string;
}